+++ CHANGELOG.md
@@ -70,6 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `--with-slurm` option is provided explicitely.
 - connect: Emit warning log entry when slurmrestd local authentication method is
   used.
+- ldap-check: Return exit code 1 on LDAP authentication error.
 - conf:
   - Bump `[slurmrestd]` > `version` default value from `0.0.40` to `0.0.41` in
     agent configuration for compatibility with Slurm 25.05.
@@ -96,6 +97,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - frontend:
   - Rendering of nodes grid in resources page in presence of long node names.
   - Update dependencies to fix CVE-2025-46565 (vite).
+- ldap-check: Report and return error instead of crashing when LDAP URI is not
+  defined in configuration.
 
 ### Removed
 - Drop support of Slurm 23.11 and Slurm REST API v0.0.40.